Annadiendo el negocio a decimal las llaves

+++ DataMapping/Data/DT_WBS_RISK_DAMAGE.cs
@@ -22,7 +22,7 @@ namespace DataMapping.Data
         {
             DataColumn[] PK_COLUMN = new DataColumn[3];
             this.TableName = TABLE_NAME;
-            PK_COLUMN[0] = this.Columns.Add(ID_WBS, typeof(int));
+            PK_COLUMN[0] = this.Columns.Add(ID_WBS, typeof(decimal));
             PK_COLUMN[1] = this.Columns.Add(ID_DAMAGE, typeof(decimal));
             PK_COLUMN[2] = this.Columns.Add(ID_RISK, typeof(decimal));
             Columns.Add(WBS, typeof(string));
